extract interfaces into separate file
