--- Tasks/UsePythonVersionV0/TaskParameters.ts
@@ -1,6 +1,0 @@
-export default interface TaskParameters {
-  versionSpec: string,
-  allowUnstable?: boolean,
-  addToPath: boolean,
-  architecture: string
-}
